Move AES implementation to toy cipher module
